fix some errors add starter for chia service add auto add directory to with plots to harvester (optional) add pause before thread start

+++ main.py
@@ -2,24 +2,24 @@ import atexit
 import configparser
 import os
 import re
+import signal
 import subprocess
 import time
 from datetime import datetime
 from pathlib import Path
-from select import select
 from threading import Thread
 
 import psutil as psutil
-from flask import Flask, request, send_from_directory, abort
+from flask import Flask, request, send_from_directory, abort, Response
 from psutil import NoSuchProcess
 
 
 def find_chia():
-    for path in Path('/').rglob('chia.exe'):
-        if 20_000_000 > path.stat().st_size > 4_000_000:
-            return path
+    for found_path in Path('/').rglob('chia.exe'):
+        if 20_000_000 > found_path.stat().st_size > 4_000_000:
+            return found_path
         else:
-            print(f'PATH FOUND: {path} {path.stat().st_size:3d}')
+            print(f'PATH FOUND: {found_path} {found_path.stat().st_size:3d}')
 
 
 def read_params_from_section(config_, section, default=None):
@@ -59,9 +59,9 @@ def convert_param_to_attribute(key, value):
     return ''
 
 
-def get_command_args(congig_dict):
-    args = [convert_param_to_attribute(key, val) for key, val in congig_dict.items()]
-    command = congig_dict['chia_path']
+def get_command_args(config_dict):
+    args = [convert_param_to_attribute(key, val) for key, val in config_dict.items()]
+    command = config_dict['chia_path']
     return f'{command} plots create {" ".join(args)}'
 
 
@@ -74,28 +74,38 @@ def get_threads_configs():
                       'work_dir': '/safe', 'memory': '4608',
                       'bucket': '128', 'k_size': '32',
                       'plots_count': 10,
+                      'auto_restart': False,
+                      'auto_find_exe': True,
+                      'pause_before_start': 0,
+                      'recheck_work_dir': False,
+                      'start_node': None,
+                      # 'set_peer_address': '',
                       'bitfield_disable': 'False'}
     if 'default' in config.sections():
         default_config = read_params_from_section(config, 'default', default_config)
     if 'chia_path' not in default_config or not Path(default_config['chia_path']).exists():
-        path = find_chia()
-        if path:
-            default_config['chia_path'] = path
+        if default_config.get('auto_find_exe', True):
+            path_chia_exe = find_chia()
+        else:
+            path_chia_exe = False
+
+        if path_chia_exe:
+            default_config['chia_path'] = path_chia_exe
             print('WARNING: NO "chia_path" IN CONFIG FILE OR PATH INCORRECT!')
-            print(f'USE PATH: {path}')
+            print(f'USE PATH: {path_chia_exe}')
         else:
             print('ERROR: NO CHIA.EXE FILE FOUND!')
             exit(1)
     config_thread = [read_params_from_section(config, section, default_config) for section in config.sections()
                      if section != 'default']
-    return config_thread
+    return config_thread, default_config
 
 
 matching = re.compile(r'Starting phase ([0-9]*/[0-9]*)')
 
 
 class ChieThread(Thread):
-    def __init__(self, name, file, cmd, current, last, temp_dir):
+    def __init__(self, name, file, cmd, current, last, temp_dir, config_for_thread):
         super().__init__(name=name)
         self.file = file
         self.cmd = cmd
@@ -107,6 +117,7 @@ class ChieThread(Thread):
         self.log = open(f'{self.name}.log', 'at')
         self.phase = 'init'
         self.last_time = 'unknown'
+        self.config = config_for_thread
 
     def __del__(self):
         if self.log:
@@ -119,67 +130,85 @@ class ChieThread(Thread):
         with open(self.file, 'wt') as file:
             file.write(f'{self.current}')
 
+    def write_log(self, text, need_time_stamp=True, need_stdout=False, need_flush=False):
+        date = datetime.now().strftime("%d.%m.%Y %H:%M:%S ") if need_time_stamp else ''
+        if need_stdout:
+            print(f'{date}{text}')
+        if self.log:
+            try:
+                self.log.write(f'{date}{text}\n')
+                if need_flush and self.log:
+                    self.log.flush()
+            except IOError:
+                pass
+
     def clear_temp(self):
-        self.log.write(
-            f'{datetime.now().strftime("%d.%m.%Y %H:%M:%S")} CLEAR TEMP\n')
+        self.write_log('CLEAR TEMP')
         files = Path(self.temp_dir).glob('*.tmp')
         for f in files:
             try:
                 f.unlink()
             except OSError as e:
                 print("Error: %s : %s" % (f, e.strerror))
-                current_date = datetime.now()
-                self.log.write(
-                    f'{current_date.strftime("%d.%m.%Y %H:%M:%S")} ERROR in {f} - {e.strerror} \n\n')
+                self.write_log(f'ERROR in {f} - {e.strerror} ')
 
     def run(self) -> None:
         self.clear_temp()
-        self.log.write(
-            f'{datetime.now().strftime("%d.%m.%Y %H:%M:%S")} START WORK from {self.current} to {self.last}\n')
+        pause = self.config.get('pause_before_start', 0)
+        if pause:
+            try:
+                pause = float(pause)
+                if pause > 0:
+                    time.sleep(pause)
+            except ValueError:
+                pass
+
+        self.write_log(f'START WORK from {self.current} to {self.last}')
         for ind in range(self.current, self.last):
             try:
                 start_time = datetime.now()
-                self.process = subprocess.Popen(['powershell', self.cmd], stderr=subprocess.PIPE,
-                                                stdout=subprocess.PIPE, shell=True, encoding='cp866',
-                                                universal_newlines=True)
+                if not self.need_stop:
+                    self.process = subprocess.Popen(['powershell', self.cmd], stderr=subprocess.PIPE,
+                                                    stdout=subprocess.PIPE, shell=True, encoding='cp866',
+                                                    universal_newlines=True)
                 while not self.need_stop and self.process.poll() is None:
                     text = self.process.stdout.readline()
-                    print(text)
                     result = matching.search(text)
                     if result:
-                        print(f'THREAD {self.name} IS {result}')
                         self.phase = result.group(1)
-                    self.log.write(text)
-                    self.log.flush()
+                    self.write_log(text, False, need_flush=True)
+                if self.need_stop and self.process.poll() is None:
+                    self.kill()
 
-                if self.process.returncode == 0:
-                    time = datetime.now() - start_time
-                    print(f'COMMAND {self.cmd}')
-                    # process.wait()
+                if self.process and self.process.returncode == 0:
+                    elapsed_time = datetime.now() - start_time
                     self.current = ind + 1
                     self.write_last()
-                    current_date = datetime.now()
-                    self.log.write(
-                        f'{current_date.strftime("%d.%m.%Y %H:%M:%S")} plot {self.current:3d} created at {time}\n')
+                    self.write_log(f'plot {self.current:3d} created at {elapsed_time}\n')
                 elif not self.need_stop:
-                    current_date = datetime.now()
                     error_text = self.process.stderr.read()
-                    self.log.write(
-                        f'{current_date.strftime("%d.%m.%Y %H:%M:%S")} ERROR {self.current:3d}: {error_text}\n')
+                    self.write_log(f'ERROR {self.current:3d}: {error_text}\n')
 
                 if self.need_stop:
-                    current_date = datetime.now()
-                    self.log.write(
-                        f'{current_date.strftime("%d.%m.%Y %H:%M:%S")} ABORT plot {self.current} \n\n')
+                    self.write_log(f'ABORT plot {self.current} \n\n', need_flush=True, need_stdout=True)
                     self.process = None
                     return
                 self.process = None
-                self.log.flush()
             except Exception as e:
                 self.process = None
-                current_date = datetime.now()
-                self.log.write(
-                    f'{current_date.strftime("%d.%m.%Y %H:%M:%S")} ERROR {e} plot {self.current} created \n\n')
+                self.write_log(f'ERROR {e} plot {self.current} created \n\n')
+
+            if self.config.get('recheck_work_dir', False):
+                chia_cmd_path = self.config.get('chia_path')
+                if chia_cmd_path:
+                    work_dir = self.config['work_dir']
+                    subprocess.Popen(['powershell', f'{chia_cmd_path} --final_dir "{work_dir}"'])
+
+        self.process = None
+        self.need_stop = True
+        self.phase = 'EXECUTED'
+        self.log.close()
+        self.log = None
 
     def kill(self):
         self.need_stop = True
@@ -187,11 +216,16 @@ class ChieThread(Thread):
             try:
                 process = psutil.Process(self.process.pid)
                 for proc in process.children(recursive=True):
-                    proc.kill()
+                    try:
+                        proc.kill()
+                        os.kill(proc.pid, signal.SIGTERM)
+                    except Exception:
+                        pass
             except NoSuchProcess:
                 pass
             time.sleep(1)
-            self.process.kill()
+            if self.process:
+                self.process.kill()
             self.process = None
         time.sleep(2)
         self.clear_temp()
@@ -209,9 +243,9 @@ class ChieThreadConfig:
         self.process = [self.get_created_plots(num) for num in range(self.num_parallel)]
 
     def get_created_plots(self, number):
-        path = Path(f'{self.name}_{number}')
-        if path.exists():
-            with open(path, 'rt') as file:
+        path_safe_iteration_count_file = Path(f'{self.name}_{number}')
+        if path_safe_iteration_count_file.exists():
+            with open(path_safe_iteration_count_file, 'rt') as file:
                 created = file.readline()
             if created.isdigit():
                 with open(f'{self.name}.log', 'at') as log:
@@ -225,10 +259,12 @@ class ChieThreadConfig:
         return 0
 
     def get_threads(self):
-        return [ChieThread(f'{self.name}-{index}', f'{self.name}_{index}', self.command, number, self.num_plots,
-                           self.config['temp_dir']) for
-                index, number in
-                enumerate(self.process)]
+        result = [ChieThread(f'{self.name}-{index}', f'{self.name}_{index}', self.command, number, self.num_plots,
+                             self.config['temp_dir'], self.config) for
+                  index, number in
+                  enumerate(self.process)]
+        self.process = [0 for _ in range(self.num_parallel)]
+        return result
 
 
 app = Flask(__name__)
@@ -238,38 +274,60 @@ class MainThread(Thread):
 
     def __init__(self):
         super().__init__()
-        self.configs = get_threads_configs()
+        self.configs, self.main_config = get_threads_configs()
         self.threads = []
-        self.init_thread()
+        self.all_thread_stopped = True
+        self.web_server_running = True
+        self.restart_command = False
+
+    def need_start(self):
+        if not self.threads:
+            return True
+        if self.all_thread_stopped and self.main_config.get('auto_restart', False):
+            return True
+        if self.all_thread_stopped and self.restart_command:
+            self.restart_command = False
+            return True
+        return False
 
     def init_thread(self):
         self.threads = [thread for conf in self.configs for thread in ChieThreadConfig(conf).get_threads()]
 
     def start_workers(self):
+        self.all_thread_stopped = False
         for thread in self.threads:
             thread.start()
 
         for thread in self.threads:
             thread.join()
+        self.all_thread_stopped = True
 
     def run(self):
-        app.run(port=5050, host='0.0.0.0')
+        try:
+            app.run(port=5050, host='0.0.0.0')
+        except Exception as _:
+            pass
+        self.web_server_running = False
 
     def kill_all(self):
+        self.main_config['auto_restart'] = False
         print('!TRY KILL ALL!')
         for thread in self.threads:
             thread.kill()
         self.shutdown_server()
 
-    def shutdown_server(self):
+    @staticmethod
+    def shutdown_server():
         try:
             func = request.environ.get('werkzeug.server.shutdown')
             if func:
                 func()
         except RuntimeError:
             pass
+        os.kill(os.getpid(), signal.CTRL_C_EVENT)
 
 
+GIGABYTE = 1024 * 1024 * 1024
 if __name__ == '__main__':
     processor = MainThread()
     atexit.register(processor.kill_all)
@@ -277,6 +335,9 @@ if __name__ == '__main__':
 
     @app.route('/')
     def hello_world():
+        load_avg = psutil.getloadavg()
+        cpu_percent = psutil.cpu_percent()
+        memory = psutil.virtual_memory()
         context = '\n'.join(
             [f'<li>{index:2d}.ПОТОК {thread.name} ВЫПОЛЯЕТ {thread.current} из {thread.last} <BR>\
                 ТЕКУЩАЯ ФАЗА {thread.phase} ВРЕМЯ ПОСЛЕДНЕГО ПЛОТА В ПОТОКЕ {thread.last_time}<BR>\
@@ -285,15 +346,26 @@ if __name__ == '__main__':
                 <a href="/stop{index}">ОСТАНОВИТЬ ПОТОК</a> \
                 </li>' for index, thread in
              enumerate(processor.threads)])
-        return f'<HTML><HEAD></HEAD><BODY><a href="/">ОБНОВИТЬ</a><UL>{context}</UL><div><a href="/stop_all">SHUTDOWN ALL</a></div></BODY></HTML>'
+        return f'<HTML><HEAD></HEAD> \
+                 <BODY> \
+                 <div>ЗАГРУЗКА CPU {cpu_percent}% \
+                 последние 1 мин {load_avg[0]}% 5 мин {load_avg[1]}% 15 мин {load_avg[2]}% \
+                 </div> \
+                 <div>\
+                 Использование памяти доступно {(memory.available / GIGABYTE):6.2f}Gb \
+                 из {(memory.total / GIGABYTE):6.2f}Gb\
+                 </div>\
+                 <a href="/">ОБНОВИТЬ</a> \
+                 <a href="restart_workers">RESTART</a> \
+                 <UL>{context}</UL><div><a href="/stop_all">SHUTDOWN</a></div></BODY></HTML>'
 
 
     @app.route('/log<name>')
     def get_log(name):
         current_path = os.getcwd()
-        path = Path(current_path).joinpath(f'{name}.log')
-        print(path)
-        if path.exists():
+        path_to_log_file = Path(current_path).joinpath(f'{name}.log')
+        print(path_to_log_file)
+        if path_to_log_file.exists():
             return send_from_directory(directory=current_path, filename=f'{name}.log')
         else:
             abort(404)
@@ -307,23 +379,31 @@ if __name__ == '__main__':
         if 0 <= index < len(processor.threads):
             processor.threads[index].kill()
             context = f'THREAD STOPPED!<BR><a href="/">BACK</a>'
-            return f'<HTML><HEAD></HEAD><BODY><UL>{context}</UL><div><a href="/stop_all">SHUTDOWN ALL</a></div></BODY></HTML>'
+            return f'<HTML><HEAD></HEAD><BODY><UL>{context}</UL></BODY></HTML>'
         else:
             abort(404)
+
+
+    def line_by_line(file_path):
+        if Path(file_path).exists():
+            with open(file_path, 'rt') as file:
+                for row in file:
+                    yield f'{row}\n'
+        else:
+            raise Exception(f'FILE {file_path} not found')
 
 
     @app.route('/view_log/<name>')
     def view_log(name):
         current_path = os.getcwd()
-        path = Path(current_path).joinpath(f'{name}.log')
-        print(path)
-        if path.exists():
-            try:
-                content = open(path, 'rt').read()
-            except Exception as e:
-                print(f'EXCEPTION {e}')
-                abort(500)
-            return content
+        path_to_log_file = Path(current_path).joinpath(f'{name}.log')
+        if path_to_log_file.exists():
+            # try:
+            #     content = open(path, 'rt').read()
+            # except Exception as e:
+            #     print(f'EXCEPTION {e}')
+            #     abort(500)
+            return Response(line_by_line(path_to_log_file), mimetype='text/csv')
         else:
             abort(404)
 
@@ -331,8 +411,32 @@ if __name__ == '__main__':
     @app.route('/stop_all')
     def stop_all():
         processor.kill_all()
-        return 'TRY ALL STOP'
+        context = 'КОМАНДА ОСТАНОВКИ ПРОГРАММЫ (завершение в течении 1 минуту)'
+        return f'<HTML><HEAD></HEAD><BODY><UL>{context}</UL></BODY></HTML>'
 
+
+    @app.route('/restart_workers')
+    def restart_all():
+        processor.restart_command = True
+        context = f'TRY RESTART!<BR><a href="/">BACK</a>'
+        return f'<HTML><HEAD></HEAD><BODY><UL>{context}</UL></BODY></HTML>'
+
+    node_start = processor.main_config.get('start_node', None)
+    if node_start:
+        path = processor.main_config['chia_path']
+        if path:
+            peer_config = processor.main_config.get('set_peer_address', None)
+            shel = subprocess.Popen(['powershell', f'{path} configure --set-farmer-peer {node_start} -upnp false'])
+            shel.wait(10)
+            subprocess.Popen(['powershell', f'{path} start {peer_config}'])
 
     processor.start()
-    processor.start_workers()
+    while processor.web_server_running:
+        if processor.need_start():
+            processor.init_thread()
+            processor.start_workers()
+        else:
+            try:
+                time.sleep(60)
+            except KeyboardInterrupt:
+                processor.kill_all()
